Add support for DISABLE_NOTIFICATIONS param
We need to be able to configure the poller to run in a "dry run" mode
where it connects to Sierra, identifies holdshelf events, logs the
number found, but doesn't attempt to hit the PatronServices Notify
endpoint or update Redis. This will aid connection debugging, allow us
to deploy and run the poller regardless of PatronServices Notify app
state, and give us a kill switch should we need to disable
notifications.