Fixes window of polling to accommodate retrieval
Fixes bug where the polling window of 20 minutes fails to catch
holdshelf evets because we're querying based on placed_gmt, which may,
in the worst case, be many hours before the item is placed on holdshelf.
This issue was written up extensively in the TAD that I wrote but did
not read.